Interfaz
Modificacion de la interfaz y funcionamiento basado en la interfaz asignada
