+++ src/pages/Mostar_video/crear-video.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { IonicPageModule } from 'ionic-angular';
+import { CrearVideoPage } from './crear-video';
+
+@NgModule({
+  declarations: [
+    CrearVideoPage,
+  ],
+  imports: [
+    IonicPageModule.forChild(CrearVideoPage),
+  ],
+})
+export class CrearVideoPageModule {}
