--- src/pages/crear-video/crear-video.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { IonicPageModule } from 'ionic-angular';
-import { CrearVideoPage } from './crear-video';
-
-@NgModule({
-  declarations: [
-    CrearVideoPage,
-  ],
-  imports: [
-    IonicPageModule.forChild(CrearVideoPage),
-  ],
-})
-export class CrearVideoPageModule {}
