--- src/pages/crear-video/crear-video.scss
@@ -1,3 +1,0 @@
-page-crear-video {
-
-}
