+++ src/pages/home/home.html
@@ -20,6 +20,7 @@
         </button>
      </ion-list>
   
-<button ion-button color="light" (click)="GoIntro()">Cartas </button>
-<button ion-button color="light" (click)="GoMapa()">Mapa </button>
+<button ion-button color="light" (click)="GoIntro()"><ion-icon name="albums"></ion-icon> </button>
+
+<button ion-button color="light" (click)="GoMapa()"> <ion-icon  name="map"></ion-icon> </button>
 </ion-content>
