+++ src/pages/intro/intro.html
@@ -28,7 +28,7 @@
       </ion-col>
       <ion-col text-right>
         <button ion-button clear small color="danger" icon-start>
-          <ion-icon name='share-alt'></ion-icon>
+            <ion-icon name="share"></ion-icon>
           Share
    
         </button>
